+++ src/main/kotlin/uk/co/ben_gibson/git/link/Context.kt
@@ -8,5 +8,4 @@ sealed class Context(val file: VirtualFile)
 
 class ContextCommit(file: VirtualFile, val commit: Commit) : Context(file)
 class ContextFileAtCommit(file: VirtualFile, val commit: Commit, val lineSelection: LineSelection? = null) : Context(file)
-class ContextFileAtBranch(file: VirtualFile, val branch: String, val lineSelection: LineSelection? = null) : Context(file)
 class ContextCurrentFile(file: VirtualFile, val lineSelection: LineSelection? = null) : Context(file)