+++ src/main/kotlin/uk/co/ben_gibson/git/link/pipeline/TimerMiddleware.kt
@@ -12,13 +12,19 @@ class TimerMiddleware : Middleware {
     override val priority = 40
 
     override fun invoke(pass: Pass, next: () -> URL?) : URL? {
+        val settings = pass.project.service<ProjectSettings>()
+
+        if (!settings.checkCommitOnRemote || !settings.showPerformanceTip) {
+            return next()
+        }
+
         val startTime = System.currentTimeMillis()
 
         val url = next()
 
         val total = System.currentTimeMillis() - startTime;
 
-        if (total > 1000 && pass.project.service<ProjectSettings>().checkCommitOnRemote) {
+        if (total > 1000) {
             sendNotification(Notification.performanceTips(pass.project), pass.project)
         }